+++ tests/step_test.py
@@ -1,7 +1,6 @@
 import pytest
 
 from tango.step import Step
-from tango.step_cache import LocalStepCache
 from tango.common.testing import TangoTestCase
 from tango.common.from_params import FromParams
 
@@ -16,13 +15,9 @@ class TestStep(TangoTestCase):
             def run(self, result: float) -> float:  # type: ignore
                 return result
 
-    def setup_method(self):
-        super().setup_method()
-        self.step_cache = LocalStepCache(self.TEST_DIR / "step_cache")
-
     def test_from_params(self):
         step = Step.from_params({"type": "float", "result": 3})
-        result = step.result(self.step_cache)
+        result = step.run_with_work_dir(self.TEST_DIR / "work")
         assert result == 3
 
     def test_from_params_wrong_type(self):
@@ -40,4 +35,4 @@ class TestStep(TangoTestCase):
                 return bar
 
         step = Step.from_params({"type": "foo", "bar": {"x": 1}})
-        assert step.result(self.step_cache).x == 1
+        assert step.run_with_work_dir(self.TEST_DIR / "work").x == 1